requirements.txt
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.14.27
botocore==1.17.27
branca==0.4.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.0
docutils==0.15.2
Flask==1.1.2
Flask-S3==0.3.3
folium==0.11.0
geographiclib==1.50
geopy==2.0.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.0.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
urllib3==1.25.10
Werkzeug==0.16.1
xlrd==1.2.0
